chore : base package scan 설정 삭제를 위한 패키지명 변경
