+++ data/user-active-rdb/src/main/java/jayho/oneday/repository/ArticleRepository.java
@@ -1,0 +1,15 @@
+package jayho.oneday.repository;
+
+import jayho.oneday.entity.Article;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+
+public interface ArticleRepository extends JpaRepository<Article, Long> {
+
+    Optional<Article> findById(Long articleId);
+//    Article findArticleResponseById(Long articleId);
+//    List<Article> findAll();
+//    List<ArticleResponseData> findArticleResponseByPage(Integer pageSize,  Long lastArticleId);
+}
