+++ data/user-active-rdb/src/main/java/jayho/oneday/repository/SavedArticleRepository.java
@@ -1,0 +1,11 @@
+package jayho.oneday.repository;
+
+
+
+import jayho.oneday.entity.SavedArticle;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+
+public interface SavedArticleRepository extends JpaRepository<SavedArticle, Long> {
+
+}
