+++ data/user-active-rdb/src/main/java/jayho/oneday/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package jayho.oneday.repository;
+
+
+import jayho.oneday.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findById(Long userId);
+}
