--- data/user-active-rdb/src/main/java/jayho/useractiverdb/repository/UserRepository.java
@@ -1,14 +1,0 @@
-package jayho.useractiverdb.repository;
-
-
-import jayho.useractiverdb.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Optional;
-
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findById(Long userId);
-}
