+++ server/user-server/src/main/java/jayho/oneday/UserServerApplication.java
@@ -1,0 +1,14 @@
+package jayho.oneday;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+//@SpringBootApplication
+@SpringBootApplication
+public class UserServerApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(UserServerApplication.class, args);
+    }
+
+}
