+++ server/user-server/src/main/java/jayho/oneday/controller/exception/ApiException.java
@@ -1,0 +1,12 @@
+package jayho.oneday.controller.exception;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class ApiException extends RuntimeException{
+
+    private final ExceptionInfo exceptionInfo;
+
+}
