+++ server/user-server/src/main/java/jayho/oneday/controller/exception/ExceptionInfo.java
@@ -1,0 +1,11 @@
+package jayho.oneday.controller.exception;
+
+import org.springframework.http.HttpStatus;
+
+public interface ExceptionInfo {
+
+    String name();
+    HttpStatus getStatus();
+    String getMessage();
+
+}
