+++ server/user-server/src/main/java/jayho/oneday/service/request/ArticleSaveRequest.java
@@ -1,0 +1,12 @@
+package jayho.oneday.service.request;
+
+import jakarta.validation.constraints.NotNull;
+import lombok.Getter;
+
+@Getter
+public class ArticleSaveRequest {
+
+    @NotNull
+    private Long userId;
+
+}
