+++ server/user-server/src/main/java/jayho/oneday/service/request/CommentUpdateRequest.java
@@ -1,0 +1,18 @@
+package jayho.oneday.service.request;
+
+import jakarta.validation.constraints.NotNull;
+import lombok.Getter;
+
+@Getter
+public class CommentUpdateRequest {
+
+    @NotNull
+    private Long articleId;
+
+    @NotNull
+    private String content;
+
+    @NotNull
+    private Long writerId;
+
+}
