+++ server/user-server/src/main/java/jayho/oneday/service/request/LikeRequest.java
@@ -1,0 +1,10 @@
+package jayho.oneday.service.request;
+
+import lombok.Getter;
+
+@Getter
+public class LikeRequest {
+
+    private Long userId;
+
+}
