+++ server/user-server/src/main/java/jayho/oneday/service/response/ArticleResponseData.java
@@ -1,0 +1,36 @@
+package jayho.oneday.service.response;
+
+import jayho.oneday.entity.Article;
+import jayho.oneday.entity.User;
+import lombok.*;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+@Getter
+@ToString
+public class ArticleResponseData {
+
+    private Long articleId;
+    private List<String> images;
+    private String content;
+    private Long writerId;
+    private String writerName;
+    private String writerProfileImage;
+    private LocalDateTime createdAt;
+    private LocalDateTime modifiedAt;
+
+
+    public static ArticleResponseData from(Article article, User user) {
+        ArticleResponseData data = new ArticleResponseData();
+        data.articleId = article.getArticleId();
+        data.images = article.getImages();
+        data.content = article.getContent();
+        data.writerId = user.getUserId();
+        data.writerName = user.getUserName();
+        data.writerProfileImage = user.getUserProfileImage();
+        data.createdAt = article.getCreatedAt();
+        data.modifiedAt = article.getModifiedAt();
+        return data;
+    }
+}
