+++ server/user-server/src/main/java/jayho/oneday/service/response/CommentResponseData.java
@@ -1,0 +1,35 @@
+package jayho.oneday.service.response;
+
+
+import jayho.oneday.entity.Comment;
+import jayho.oneday.entity.User;
+import lombok.Getter;
+import lombok.ToString;
+
+@Getter
+@ToString
+public class CommentResponseData {
+
+    private Long commentId;
+    private String content;
+    private Long writerId;
+    private String writerName;
+    private String writerProfileImage;
+    private Long articleId;
+
+    public static CommentResponseData from(Comment comment, User writer) {
+
+        CommentResponseData response = new CommentResponseData();
+        response.commentId = comment.getCommentId();
+        response.content = comment.getContent();
+        response.writerId = comment.getWriterId();
+        response.writerName = writer.getUserName();
+        response.writerProfileImage = writer.getUserProfileImage();
+        response.articleId = comment.getArticleId();
+
+        return response;
+    }
+
+
+
+}
