--- server/user-server/src/main/java/jayho/userserver/UserServerApplication.java
@@ -1,14 +1,0 @@
-package jayho.userserver;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication(scanBasePackages = {"jayho.userserver", "jayho.useractiverdb"})
-public class UserServerApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(UserServerApplication.class, args);
-    }
-
-
-}
