--- server/user-server/src/main/java/jayho/userserver/controller/exception/ApiException.java
@@ -1,12 +1,0 @@
-package jayho.userserver.controller.exception;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class ApiException extends RuntimeException{
-
-    private final ExceptionInfo exceptionInfo;
-
-}
