--- server/user-server/src/main/java/jayho/userserver/controller/exception/ExceptionInfo.java
@@ -1,11 +1,0 @@
-package jayho.userserver.controller.exception;
-
-import org.springframework.http.HttpStatus;
-
-public interface ExceptionInfo {
-
-    String name();
-    HttpStatus getStatus();
-    String getMessage();
-
-}
