--- server/user-server/src/main/java/jayho/userserver/service/request/ArticleCreateRequest.java
@@ -1,22 +1,0 @@
-package jayho.userserver.service.request;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.util.List;
-
-@Getter
-@AllArgsConstructor
-public class ArticleCreateRequest {
-
-    @NotNull
-    private List<String> images;
-    private String content;
-
-    @NotNull
-    private Long writerId;
-
-}
