--- server/user-server/src/main/java/jayho/userserver/service/request/ArticleSaveRequest.java
@@ -1,12 +1,0 @@
-package jayho.userserver.service.request;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.Getter;
-
-@Getter
-public class ArticleSaveRequest {
-
-    @NotNull
-    private Long userId;
-
-}
