--- server/user-server/src/main/java/jayho/userserver/service/request/CommentCreateRequest.java
@@ -1,21 +1,0 @@
-package jayho.userserver.service.request;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class CommentCreateRequest {
-
-    @NotNull
-    private Long articleId;
-
-    @NotNull
-    private String content;
-
-    @NotNull
-    private Long writerId;
-
-
-}
