--- server/user-server/src/main/java/jayho/userserver/service/request/CommentUpdateRequest.java
@@ -1,18 +1,0 @@
-package jayho.userserver.service.request;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.Getter;
-
-@Getter
-public class CommentUpdateRequest {
-
-    @NotNull
-    private Long articleId;
-
-    @NotNull
-    private String content;
-
-    @NotNull
-    private Long writerId;
-
-}
