--- server/user-server/src/main/java/jayho/userserver/service/request/LikeRequest.java
@@ -1,10 +1,0 @@
-package jayho.userserver.service.request;
-
-import lombok.Getter;
-
-@Getter
-public class LikeRequest {
-
-    private Long userId;
-
-}
